Reverted Passive Attributes Change
Removed the change that moved the loadout and taunt buttons to the side to make room for the passive attributes label that was added back. However this change had to be removed because the position of the buttons is influenced by the resolution and aspect ration - so unlike the other elements on the page moving them anywhere else would only work in single resolution and others they would be moved offscreen.