HUD Element movement
This updated HUD Layout moves the Demoman's Stickybomb Counter to be on the bottom of the screen as well as stretched to accommodate 10+ stickies. moves the disguise panel to be above health to not overlap anything, and moves  the Engineer's Metal Counter to be in line with the Stickybomb Counter.